requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pip==26.0.1
pycryptodome==3.23.0
Pygments==2.20.0
random-user-agent==1.0.1
requests==2.33.1
rich==14.3.3
urllib3==2.6.3
Werkzeug==3.1.3